feat: Image From   * Got keg-core working   * Added image from to the injected template     * Uses the built remote image     * Cleaned up the start service     * Added __injected check to imgNameContext, for injected apps
