+++ src/templates/injected-compose.template.yml
@@ -3,8 +3,8 @@ networks:
   {{ dockerNetwork }}:
     external: true
 services:
-  {{ image }}:
-    image: {{ image }}
+  {{ service }}:
+    image: {{ imageFrom }}
     container_name: {{ container }}
     tty: true
     stdin_open: true
